Refactor the per-cloud and per-profile code
Rather than copy-pasting around same tasks but changing slightly
the way we loop them when it's a per-profile or a per-cloud resource,
refactor the code a bit so both mechanisms call the same task.
e.g. the per-profile projects and per-cloud projects will call in the
end create_project, the only thing it varies is the loop prior to getting
the individual project items.

Depends-On: I367c291f8ef052eb285fa803bd8d8737058eef46
Depends-On: Ie96940718e6ba1e8c287c06b8fcd417ff0ed79e7
Change-Id: Ida2aebadafd4d60d11fa18aa2208dad235badd4c