Refactor the per-cloud and per-profile code
Rather than copy-pasting around same tasks but changing slightly
the way we loop them when it's a per-profile or a per-cloud resource,
refactor the code a bit so both mechanisms call the same task.
e.g. the per-profile projects and per-cloud projects will call in the
end create_project, the only thing it varies is the loop prior to getting
the individual project items.

Depends-On: I0e98066a85df21675882bd0b6b0547f29da7387a
Depends-On: I57faa3835bcf049a6a461e2a5c43381fdd884424
Change-Id: Ida2aebadafd4d60d11fa18aa2208dad235badd4c