Add ability to download an image from the Internet to upload it to Glance
Unfortunately, it seems Glance v2 does not have the capability to upload
an image by referencing a URL, it has to be a local file.
Adding logic to the role and flags should the user want to pull an image
to the Ansible control machine prior to the usual upload.

Change-Id: I2e2886287a045dab89fe712e39b42c2383d6aec6