Ability to run the launcher against a particular cloud
When we feed a resources.yaml file containing the 'clouds' dict
containing the definitions, the role will happily process all of them.
This is good, but it would also be nice to specify on the command line
that we only want to run the launcher against a specific cloud.
With this change, we can just pass -e "cloud=foocloud" to ansible-playbook
and the launcher will just process the resources for cloud 'foocloud'.

Change-Id: I37cddfbbbb83203ed62434dccac60511f726c6a4